Move additional worker-thread only functions from `src/shared/util.js` and into a `src/core/core_utils.js` instead
This moves the `log2`, `readInt8`, `readUint16`, `readUint32`, and `isSpace` functions since they are only used in the worker-thread.
